Allow building a non-PCL and put PCL and non-PCL in nupkg.
Passing /p:PCL=false to msbuild will build a non-PCL dll. This
is necessary for C5 to run on versions of Mono earlier than 3.x.

I'm not sure what effect the conditionals inserted into the
.csproj have on Visual Studio. I am not able to open C5 in
Visual Studio because I only have the Express edition.

Changed nuspec to contain id, version, etc directly because
the package will need to be built passing the .nuspec instead of
the .csproj in order to build a package with multiple target
frameworks (net40 and portable-win+net40+sl40+wp+Xbox40).

Added a .bat script for building a package. It does a rebuild of
C5 and copies the dll and xml file to
package\lib\portable-win+net40+sl40+wp+Xbox40\

then does it again, passing PCL=false to msbuild and copying
to package\lib\net40\

Finally it does a nuget pack.